renamed types to util/constants
