+++ src/components/pages/gallery/PreviewCard.tsx
@@ -5,7 +5,7 @@ import PlayCircleOutline from 'components/icons/PlayCircleOutline';
 import DownloadManager from 'services/downloadManager';
 import useLongPress from 'utils/common/useLongPress';
 import { GalleryContext } from 'pages/gallery';
-import { GAP_BTW_TILES } from 'types';
+import { GAP_BTW_TILES } from 'utils/constants';
 
 interface IProps {
     file: EnteFile;
