+++ src/pages/change-email/index.tsx
@@ -8,7 +8,7 @@ import { getToken } from 'utils/common/key';
 import EnteSpinner from 'components/EnteSpinner';
 import ChangeEmailForm from 'components/ChangeEmail';
 import EnteCard from 'components/EnteCard';
-import { PAGES } from 'types';
+import { PAGES } from 'utils/constants';
 
 function ChangeEmailPage() {
     const [email, setEmail] = useState('');
